refactor(dom): attachComponent()
