+++ core/src/index.d.ts
@@ -107,7 +107,7 @@ export { PlatformConfig } from './global/platform-configs';
 export * from './components';
 
 export { DomController, RafCallback } from './global/dom-controller';
-export { FrameworkDelegate } from './utils/dom-framework-delegate';
+export { FrameworkDelegate } from './utils/framework-delegate';
 export { OverlayEventDetail } from './utils/overlays';
 
 export interface Config {
